https://github.com/jackdewinter/pymarkdown/issues/1482
