+++ test/test_markdown_extra.py
@@ -17335,418 +17335,6 @@ LRD.</p>"""
 
 
 @pytest.mark.gfm
-def test_extra_issue_1563_a() -> None:
-    """
-    TBD
-    """
-
-    # Arrange
-    source_markdown = """
-> a block
->
-> 1. another list
-     properly indented content
->  1. another list
-      properly indented content
-"""
-    expected_tokens = [
-        "[BLANK(1,1):]",
-        "[block-quote(2,1)::> \n>\n> \n\n> \n]",
-        "[para(2,3):]",
-        "[text(2,3):a block:]",
-        "[end-para:::True]",
-        "[BLANK(3,2):]",
-        "[olist(4,3):.:1:5::     \n      ]",
-        "[para(4,6):\n]",
-        "[text(4,6):another list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[li(6,4):6: :1]",
-        "[para(6,7):\n]",
-        "[text(6,7):another list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[end-olist:::True]",
-        "[end-block-quote:::True]",
-        "[BLANK(8,1):]",
-    ]
-    expected_gfm = """<blockquote>
-<p>a block</p>
-<ol>
-<li>another list
-properly indented content</li>
-<li>another list
-properly indented content</li>
-</ol>
-</blockquote>"""
-
-    # Act & Assert
-    act_and_assert(source_markdown, expected_gfm, expected_tokens)
-
-
-@pytest.mark.gfm
-def test_extra_issue_1563_b() -> None:
-    """
-    TBD
-    """
-
-    # Arrange
-    source_markdown = """
-> a block
->
-> 1. another list
-     properly indented content
-> 1. another list
-     properly indented content
-"""
-    expected_tokens = [
-        "[BLANK(1,1):]",
-        "[block-quote(2,1)::> \n>\n> \n\n> \n]",
-        "[para(2,3):]",
-        "[text(2,3):a block:]",
-        "[end-para:::True]",
-        "[BLANK(3,2):]",
-        "[olist(4,3):.:1:5::     \n     ]",
-        "[para(4,6):\n]",
-        "[text(4,6):another list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[li(6,3):5::1]",
-        "[para(6,6):\n]",
-        "[text(6,6):another list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[end-olist:::True]",
-        "[end-block-quote:::True]",
-        "[BLANK(8,1):]",
-    ]
-    expected_gfm = """<blockquote>
-<p>a block</p>
-<ol>
-<li>another list
-properly indented content</li>
-<li>another list
-properly indented content</li>
-</ol>
-</blockquote>"""
-
-    # Act & Assert
-    act_and_assert(source_markdown, expected_gfm, expected_tokens)
-
-
-@pytest.mark.gfm
-def test_extra_issue_1563_c() -> None:
-    """
-    TBD
-    """
-
-    # Arrange
-    source_markdown = """
-> a block
->
-> 1. another list
-     properly indented content
->  1. another list
-      properly indented content
-"""
-    expected_tokens = [
-        "[BLANK(1,1):]",
-        "[block-quote(2,1)::> \n>\n> \n\n> \n]",
-        "[para(2,3):]",
-        "[text(2,3):a block:]",
-        "[end-para:::True]",
-        "[BLANK(3,2):]",
-        "[olist(4,3):.:1:5::     \n      ]",
-        "[para(4,6):\n]",
-        "[text(4,6):another list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[li(6,4):6: :1]",
-        "[para(6,7):\n]",
-        "[text(6,7):another list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[end-olist:::True]",
-        "[end-block-quote:::True]",
-        "[BLANK(8,1):]",
-    ]
-    expected_gfm = """<blockquote>
-<p>a block</p>
-<ol>
-<li>another list
-properly indented content</li>
-<li>another list
-properly indented content</li>
-</ol>
-</blockquote>"""
-
-    # Act & Assert
-    act_and_assert(source_markdown, expected_gfm, expected_tokens)
-
-
-@pytest.mark.gfm
-def test_extra_issue_1564_a() -> None:
-    """
-    TBD
-    """
-
-    # Arrange
-    source_markdown = """
-> a block
->
-> 1. another list
- properly indented content
-> 1. another list
-     properly indented content
-"""
-    expected_tokens = [
-        "[BLANK(1,1):]",
-        "[block-quote(2,1)::> \n>\n> \n\n> \n]",
-        "[para(2,3):]",
-        "[text(2,3):a block:]",
-        "[end-para:::True]",
-        "[BLANK(3,2):]",
-        "[olist(4,3):.:1:5:: \n     ]",
-        "[para(4,6):\n]",
-        "[text(4,6):another list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[li(6,3):5::1]",
-        "[para(6,6):\n]",
-        "[text(6,6):another list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[end-olist:::True]",
-        "[end-block-quote:::True]",
-        "[BLANK(8,1):]",
-    ]
-    expected_gfm = """<blockquote>
-<p>a block</p>
-<ol>
-<li>another list
-properly indented content</li>
-<li>another list
-properly indented content</li>
-</ol>
-</blockquote>"""
-
-    # Act & Assert
-    act_and_assert(source_markdown, expected_gfm, expected_tokens)
-
-
-@pytest.mark.gfm
-def test_extra_issue_1564_b() -> None:
-    """
-    TBD
-    """
-
-    # Arrange
-    source_markdown = """
-> a block
->
-> 1. another list
-properly indented content
-> 1. another list
-     properly indented content
-"""
-    expected_tokens = [
-        "[BLANK(1,1):]",
-        "[block-quote(2,1)::> \n>\n> \n\n> \n]",
-        "[para(2,3):]",
-        "[text(2,3):a block:]",
-        "[end-para:::True]",
-        "[BLANK(3,2):]",
-        "[olist(4,3):.:1:5::\n     ]",
-        "[para(4,6):\n]",
-        "[text(4,6):another list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[li(6,3):5::1]",
-        "[para(6,6):\n]",
-        "[text(6,6):another list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[end-olist:::True]",
-        "[end-block-quote:::True]",
-        "[BLANK(8,1):]",
-    ]
-    expected_gfm = """<blockquote>
-<p>a block</p>
-<ol>
-<li>another list
-properly indented content</li>
-<li>another list
-properly indented content</li>
-</ol>
-</blockquote>"""
-
-    # Act & Assert
-    act_and_assert(
-        source_markdown,
-        expected_gfm,
-        expected_tokens,
-    )
-
-
-@pytest.mark.gfm
-def test_extra_issue_1563_d() -> None:
-    """
-    TBD
-    """
-
-    # Arrange
-    source_markdown = """
-> > a block
-> >
-> > 1. another list
-> >   improperly indented content
-> > 1. another list
-       properly indented content
-"""
-    expected_tokens = [
-        "[BLANK(1,1):]",
-        "[block-quote(2,1)::]",
-        "[block-quote(2,3)::> > \n> >\n> > \n> > \n> > \n]",
-        "[para(2,5):]",
-        "[text(2,5):a block:]",
-        "[end-para:::True]",
-        "[BLANK(3,4):]",
-        "[olist(4,5):.:1:7::  \n       ]",
-        "[para(4,8):\n]",
-        "[text(4,8):another list\nimproperly indented content::\n]",
-        "[end-para:::True]",
-        "[li(6,5):7::1]",
-        "[para(6,8):\n]",
-        "[text(6,8):another list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[end-olist:::True]",
-        "[end-block-quote:::True]",
-        "[end-block-quote:::True]",
-        "[BLANK(8,1):]",
-    ]
-    expected_gfm = """<blockquote>
-<blockquote>
-<p>a block</p>
-<ol>
-<li>another list
-improperly indented content</li>
-<li>another list
-properly indented content</li>
-</ol>
-</blockquote>
-</blockquote>"""
-
-    # Act & Assert
-    act_and_assert(
-        source_markdown,
-        expected_gfm,
-        expected_tokens,
-    )
-
-
-@pytest.mark.gfm
-def test_extra_issue_1566_a() -> None:
-    """
-    TBD
-    """
-
-    # Arrange
-    source_markdown = """
-> > a block
-> >
-> > 1. another list
-> >    properly indented content
-> > 1. another list
-> >    properly indented content
-"""
-    expected_tokens = [
-        "[BLANK(1,1):]",
-        "[block-quote(2,1)::]",
-        "[block-quote(2,3)::> > \n> >\n> > \n> > \n> > \n> > ]",
-        "[para(2,5):]",
-        "[text(2,5):a block:]",
-        "[end-para:::True]",
-        "[BLANK(3,4):]",
-        "[olist(4,5):.:1:7::   \n   ]",
-        "[para(4,8):\n]",
-        "[text(4,8):another list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[li(6,5):7::1]",
-        "[para(6,8):\n]",
-        "[text(6,8):another list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[end-olist:::True]",
-        "[end-block-quote:::True]",
-        "[end-block-quote:::True]",
-        "[BLANK(8,1):]",
-    ]
-    expected_gfm = """<blockquote>
-<blockquote>
-<p>a block</p>
-<ol>
-<li>another list
-properly indented content</li>
-<li>another list
-properly indented content</li>
-</ol>
-</blockquote>
-</blockquote>"""
-
-    # Act & Assert
-    act_and_assert(
-        source_markdown,
-        expected_gfm,
-        expected_tokens,
-    )
-
-
-@pytest.mark.gfm
-def test_extra_issue_1566_b() -> None:
-    """
-    TBD
-    """
-
-    # Arrange
-    source_markdown = """
-> > a block
-> >
-> > 1. another list
-> >    properly indented content
-> >  1. another list
-> >     properly indented content
-"""
-    expected_tokens = [
-        "[BLANK(1,1):]",
-        "[block-quote(2,1)::]",
-        "[block-quote(2,3)::> > \n> >\n> > \n> > \n> > \n> > ]",
-        "[para(2,5):]",
-        "[text(2,5):a block:]",
-        "[end-para:::True]",
-        "[BLANK(3,4):]",
-        "[olist(4,5):.:1:7::   \n    ]",
-        "[para(4,8):\n]",
-        "[text(4,8):another list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[li(6,6):8: :1]",
-        "[para(6,9):\n]",
-        "[text(6,9):another list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[end-olist:::True]",
-        "[end-block-quote:::True]",
-        "[end-block-quote:::True]",
-        "[BLANK(8,1):]",
-    ]
-    expected_gfm = """<blockquote>
-<blockquote>
-<p>a block</p>
-<ol>
-<li>another list
-properly indented content</li>
-<li>another list
-properly indented content</li>
-</ol>
-</blockquote>
-</blockquote>"""
-
-    # Act & Assert
-    act_and_assert(
-        source_markdown,
-        expected_gfm,
-        expected_tokens,
-    )
-
-
-@pytest.mark.gfm
 def test_extra_059a() -> None:
     """
     TBD
@@ -17869,494 +17457,6 @@ properly indented content</li>
 </ol>
 </blockquote>
 </blockquote>"""
-
-    # Act & Assert
-    act_and_assert(
-        source_markdown,
-        expected_gfm,
-        expected_tokens,
-    )
-
-
-@pytest.mark.gfm
-def test_extra_issue_1569_a() -> None:
-    """
-    TBD
-    """
-
-    # Arrange
-    source_markdown = """
-> 1. first list
->    1. inner list
-        properly indented content
->    1. another list
->      properly indented content
-"""
-    expected_tokens = [
-        "[BLANK(1,1):]",
-        "[block-quote(2,1)::> \n> \n\n> \n> ]",
-        "[olist(2,3):.:1:5:]",
-        "[para(2,6):]",
-        "[text(2,6):first list:]",
-        "[end-para:::True]",
-        "[olist(3,6):.:1:8:   :        \n     ]",
-        "[para(3,9):\n]",
-        "[text(3,9):inner list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[li(5,6):8:   :1]",
-        "[para(5,9):\n]",
-        "[text(5,9):another list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[end-olist:::True]",
-        "[end-olist:::True]",
-        "[end-block-quote:::True]",
-        "[BLANK(7,1):]",
-    ]
-    expected_gfm = """<blockquote>
-<ol>
-<li>first list
-<ol>
-<li>inner list
-properly indented content</li>
-<li>another list
-properly indented content</li>
-</ol>
-</li>
-</ol>
-</blockquote>"""
-
-    # Act & Assert
-    act_and_assert(
-        source_markdown,
-        expected_gfm,
-        expected_tokens,
-    )
-
-
-@pytest.mark.gfm
-def test_extra_issue_1569_b() -> None:
-    """
-    TBD
-    """
-
-    # Arrange
-    source_markdown = """
-> 1. first list
->    1. inner list
-        properly indented content
->    1. another list
->       properly indented content
-"""
-    expected_tokens = [
-        "[BLANK(1,1):]",
-        "[block-quote(2,1)::> \n> \n\n> \n> ]",
-        "[olist(2,3):.:1:5:]",
-        "[para(2,6):]",
-        "[text(2,6):first list:]",
-        "[end-para:::True]",
-        "[olist(3,6):.:1:8:   :        \n      ]",
-        "[para(3,9):\n]",
-        "[text(3,9):inner list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[li(5,6):8:   :1]",
-        "[para(5,9):\n]",
-        "[text(5,9):another list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[end-olist:::True]",
-        "[end-olist:::True]",
-        "[end-block-quote:::True]",
-        "[BLANK(7,1):]",
-    ]
-    expected_gfm = """<blockquote>
-<ol>
-<li>first list
-<ol>
-<li>inner list
-properly indented content</li>
-<li>another list
-properly indented content</li>
-</ol>
-</li>
-</ol>
-</blockquote>"""
-
-    # Act & Assert
-    act_and_assert(
-        source_markdown,
-        expected_gfm,
-        expected_tokens,
-    )
-
-
-@pytest.mark.gfm
-def test_extra_issue_1573_a() -> None:
-    """
-    TBD
-    """
-
-    # Arrange
-    source_markdown = """
-> 1. first list
->    1. inner list
-        properly indented content
->    1. another list
->        properly indented content
-"""
-    expected_tokens = [
-        "[BLANK(1,1):]",
-        "[block-quote(2,1)::> \n> \n\n> \n> ]",
-        "[olist(2,3):.:1:5:]",
-        "[para(2,6):]",
-        "[text(2,6):first list:]",
-        "[end-para:::True]",
-        "[olist(3,6):.:1:8:   :        \n      ]",
-        "[para(3,9):\n]",
-        "[text(3,9):inner list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[li(5,6):8:   :1]",
-        "[para(5,9):\n ]",
-        "[text(5,9):another list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[end-olist:::True]",
-        "[end-olist:::True]",
-        "[end-block-quote:::True]",
-        "[BLANK(7,1):]",
-    ]
-    expected_gfm = """<blockquote>
-<ol>
-<li>first list
-<ol>
-<li>inner list
-properly indented content</li>
-<li>another list
-properly indented content</li>
-</ol>
-</li>
-</ol>
-</blockquote>"""
-
-    # Act & Assert
-    act_and_assert(
-        source_markdown,
-        expected_gfm,
-        expected_tokens,
-    )
-
-
-@pytest.mark.gfm
-def test_extra_issue_1573_b() -> None:
-    """
-    TBD
-    """
-
-    # Arrange
-    source_markdown = """
-1. > first list
-   > 1. inner list
-        properly indented content
-   > 1. another list
-        properly indented content
-"""
-    expected_tokens = [
-        "[BLANK(1,1):]",
-        "[olist(2,1):.:1:3::]",
-        "[block-quote(2,4):   :   > \n   > \n\n   > \n]",
-        "[para(2,6):]",
-        "[text(2,6):first list:]",
-        "[end-para:::True]",
-        "[olist(3,6):.:1:8::        \n        ]",
-        "[para(3,9):\n]",
-        "[text(3,9):inner list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[li(5,6):8::1]",
-        "[para(5,9):\n]",
-        "[text(5,9):another list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[end-olist:::True]",
-        "[end-block-quote:::True]",
-        "[BLANK(7,1):]",
-        "[end-olist:::True]",
-    ]
-    expected_gfm = """<ol>
-<li>
-<blockquote>
-<p>first list</p>
-<ol>
-<li>inner list
-properly indented content</li>
-<li>another list
-properly indented content</li>
-</ol>
-</blockquote>
-</li>
-</ol>"""
-
-    # Act & Assert
-    act_and_assert(
-        source_markdown,
-        expected_gfm,
-        expected_tokens,
-    )
-
-
-@pytest.mark.gfm
-def test_extra_issue_1574_a() -> None:
-    """
-    TBD
-    """
-
-    # Arrange
-    source_markdown = """
-1. first list
-   1. second list
-      > inner block
-      > inner block
-      ----
-      first line
-      another list
-    1. second list
-       > inner block
-       > inner block
-       ----
-       first line
-  another fn list
-"""
-    expected_tokens = [
-        "[BLANK(1,1):]",
-        "[olist(2,1):.:1:3:]",
-        "[para(2,4):]",
-        "[text(2,4):first list:]",
-        "[end-para:::True]",
-        "[olist(3,4):.:1:6:   :\n\n      \n      \n      \n\n\n       \n       \n\n]",
-        "[para(3,7):]",
-        "[text(3,7):second list:]",
-        "[end-para:::True]",
-        "[block-quote(4,7):      :      > \n      > ]",
-        "[para(4,9):\n]",
-        "[text(4,9):inner block\ninner block::\n]",
-        "[end-para:::True]",
-        "[end-block-quote:::True]",
-        "[tbreak(6,7):-::----]",
-        "[para(7,7):\n]",
-        "[text(7,7):first line\nanother list::\n]",
-        "[end-para:::True]",
-        "[li(9,5):7:    :1]",
-        "[para(9,8):]",
-        "[text(9,8):second list:]",
-        "[end-para:::True]",
-        "[block-quote(10,8):       :       > \n       > ]",
-        "[para(10,10):\n]",
-        "[text(10,10):inner block\ninner block::\n]",
-        "[end-para:::True]",
-        "[end-block-quote:::True]",
-        "[tbreak(12,8):-::----]",
-        "[para(13,8):\n  ]",
-        "[text(13,8):first line\nanother fn list::\n]",
-        "[end-para:::True]",
-        "[BLANK(15,1):]",
-        "[end-olist:::True]",
-        "[end-olist:::True]",
-    ]
-    expected_gfm = """<ol>
-<li>first list
-<ol>
-<li>second list
-<blockquote>
-<p>inner block
-inner block</p>
-</blockquote>
-<hr />
-first line
-another list</li>
-<li>second list
-<blockquote>
-<p>inner block
-inner block</p>
-</blockquote>
-<hr />
-first line
-another fn list</li>
-</ol>
-</li>
-</ol>"""
-
-    # Act & Assert
-    act_and_assert(
-        source_markdown,
-        expected_gfm,
-        expected_tokens,
-    )
-
-
-@pytest.mark.gfm
-def test_extra_issue_1574_b() -> None:
-    """
-    TBD
-    """
-
-    # Arrange
-    source_markdown = """
-1. first list
-   1. second list
-      > inner block
-      > inner block
-      ----
-      first line
-      another list
-    1. second list
-       > inner block
-       > inner block
-       ----
-       first line
-   another fn list
-"""
-    expected_tokens = [
-        "[BLANK(1,1):]",
-        "[olist(2,1):.:1:3:]",
-        "[para(2,4):]",
-        "[text(2,4):first list:]",
-        "[end-para:::True]",
-        "[olist(3,4):.:1:6:   :\n\n      \n      \n      \n\n\n       \n       \n\n]",
-        "[para(3,7):]",
-        "[text(3,7):second list:]",
-        "[end-para:::True]",
-        "[block-quote(4,7):      :      > \n      > ]",
-        "[para(4,9):\n]",
-        "[text(4,9):inner block\ninner block::\n]",
-        "[end-para:::True]",
-        "[end-block-quote:::True]",
-        "[tbreak(6,7):-::----]",
-        "[para(7,7):\n]",
-        "[text(7,7):first line\nanother list::\n]",
-        "[end-para:::True]",
-        "[li(9,5):7:    :1]",
-        "[para(9,8):]",
-        "[text(9,8):second list:]",
-        "[end-para:::True]",
-        "[block-quote(10,8):       :       > \n       > ]",
-        "[para(10,10):\n]",
-        "[text(10,10):inner block\ninner block::\n]",
-        "[end-para:::True]",
-        "[end-block-quote:::True]",
-        "[tbreak(12,8):-::----]",
-        "[para(13,8):\n   ]",
-        "[text(13,8):first line\nanother fn list::\n]",
-        "[end-para:::True]",
-        "[BLANK(15,1):]",
-        "[end-olist:::True]",
-        "[end-olist:::True]",
-    ]
-    expected_gfm = """<ol>
-<li>first list
-<ol>
-<li>second list
-<blockquote>
-<p>inner block
-inner block</p>
-</blockquote>
-<hr />
-first line
-another list</li>
-<li>second list
-<blockquote>
-<p>inner block
-inner block</p>
-</blockquote>
-<hr />
-first line
-another fn list</li>
-</ol>
-</li>
-</ol>"""
-
-    # Act & Assert
-    act_and_assert(
-        source_markdown,
-        expected_gfm,
-        expected_tokens,
-    )
-
-
-@pytest.mark.gfm
-def test_extra_issue_1574_c() -> None:
-    """
-    TBD
-    """
-
-    # Arrange
-    source_markdown = """
-1. first list
-   1. second list
-      > inner block
-      > inner block
-      ----
-      first line
-      another list
-    1. second list
-       > inner block
-       > inner block
-       ----
-       first line
-    another fn list
-"""
-    expected_tokens = [
-        "[BLANK(1,1):]",
-        "[olist(2,1):.:1:3:]",
-        "[para(2,4):]",
-        "[text(2,4):first list:]",
-        "[end-para:::True]",
-        "[olist(3,4):.:1:6:   :\n\n      \n      \n      \n\n\n       \n       \n\n]",
-        "[para(3,7):]",
-        "[text(3,7):second list:]",
-        "[end-para:::True]",
-        "[block-quote(4,7):      :      > \n      > ]",
-        "[para(4,9):\n]",
-        "[text(4,9):inner block\ninner block::\n]",
-        "[end-para:::True]",
-        "[end-block-quote:::True]",
-        "[tbreak(6,7):-::----]",
-        "[para(7,7):\n]",
-        "[text(7,7):first line\nanother list::\n]",
-        "[end-para:::True]",
-        "[li(9,5):7:    :1]",
-        "[para(9,8):]",
-        "[text(9,8):second list:]",
-        "[end-para:::True]",
-        "[block-quote(10,8):       :       > \n       > ]",
-        "[para(10,10):\n]",
-        "[text(10,10):inner block\ninner block::\n]",
-        "[end-para:::True]",
-        "[end-block-quote:::True]",
-        "[tbreak(12,8):-::----]",
-        "[para(13,8):\n    ]",
-        "[text(13,8):first line\nanother fn list::\n]",
-        "[end-para:::True]",
-        "[BLANK(15,1):]",
-        "[end-olist:::True]",
-        "[end-olist:::True]",
-    ]
-    expected_gfm = """<ol>
-<li>first list
-<ol>
-<li>second list
-<blockquote>
-<p>inner block
-inner block</p>
-</blockquote>
-<hr />
-first line
-another list</li>
-<li>second list
-<blockquote>
-<p>inner block
-inner block</p>
-</blockquote>
-<hr />
-first line
-another fn list</li>
-</ol>
-</li>
-</ol>"""
 
     # Act & Assert
     act_and_assert(
@@ -18543,205 +17643,6 @@ another list</li>
 
 
 @pytest.mark.gfm
-def test_extra_issue_1577_a() -> None:
-    """
-    TBD
-    """
-
-    # Arrange
-    source_markdown = """
-> block start
->
-> + first list
->   > block within list
->   ----
->   first line
->   another list
->  + second list
->    > block within list
->    ----
->    first line
->    another list
-"""
-    expected_tokens = [
-        "[BLANK(1,1):]",
-        "[block-quote(2,1)::> \n>\n> \n> \n> \n> \n> \n> \n> \n> ]",
-        "[para(2,3):]",
-        "[text(2,3):block start:]",
-        "[end-para:::True]",
-        "[BLANK(3,2):]",
-        "[ulist(4,3):+::4::\n  þ\n  \n  \n\n   þ\n   \n   ]",
-        "[para(4,5):]",
-        "[text(4,5):first list:]",
-        "[end-para:::True]",
-        "[block-quote(5,5)::> \n> ]",
-        "[para(5,7):]",
-        "[text(5,7):block within list:]",
-        "[end-para:::False]",
-        "[end-block-quote::> :True]",
-        "[tbreak(6,5):-::----]",
-        "[para(7,5):\n]",
-        "[text(7,5):first line\nanother list::\n]",
-        "[end-para:::True]",
-        "[li(9,4):5: :]",
-        "[para(9,6):]",
-        "[text(9,6):second list:]",
-        "[end-para:::True]",
-        "[block-quote(10,6)::> \n> ]",
-        "[para(10,8):]",
-        "[text(10,8):block within list:]",
-        "[end-para:::False]",
-        "[end-block-quote::> :True]",
-        "[tbreak(11,6):-::----]",
-        "[para(12,6):\n]",
-        "[text(12,6):first line\nanother list::\n]",
-        "[end-para:::True]",
-        "[end-ulist:::True]",
-        "[end-block-quote:::True]",
-        "[BLANK(14,1):]",
-    ]
-    expected_gfm = """<blockquote>
-<p>block start</p>
-<ul>
-<li>first list
-<blockquote>
-<p>block within list</p>
-</blockquote>
-<hr />
-first line
-another list</li>
-<li>second list
-<blockquote>
-<p>block within list</p>
-</blockquote>
-<hr />
-first line
-another list</li>
-</ul>
-</blockquote>"""
-
-    # Act & Assert
-    act_and_assert(
-        source_markdown,
-        expected_gfm,
-        expected_tokens,
-        show_debug=False,
-    )
-
-
-@pytest.mark.gfm
-def test_extra_issue_1577_b() -> None:
-    """
-    TBD
-    """
-
-    # Arrange
-    source_markdown = """
-> block start
->
-> + first list
->   > block within list
->   ----
->   first line
->   another list
-> + second list
->   > block within list
->   ----
->   first line
->   another list
->  + third list
->    > block within list
->    ----
->    first line
->    another list
-"""
-    expected_tokens = [
-        "[BLANK(1,1):]",
-        "[block-quote(2,1)::> \n>\n> \n> \n> \n> \n> \n> \n> \n> \n> \n> \n> \n> ]",
-        "[para(2,3):]",
-        "[text(2,3):block start:]",
-        "[end-para:::True]",
-        "[BLANK(3,2):]",
-        "[ulist(4,3):+::4::\n  þ\n  \n  \n\n  þ\n  \n  \n\n   þ\n   \n   ]",
-        "[para(4,5):]",
-        "[text(4,5):first list:]",
-        "[end-para:::True]",
-        "[block-quote(5,5)::> \n> ]",
-        "[para(5,7):]",
-        "[text(5,7):block within list:]",
-        "[end-para:::False]",
-        "[end-block-quote::> :True]",
-        "[tbreak(6,5):-::----]",
-        "[para(7,5):\n]",
-        "[text(7,5):first line\nanother list::\n]",
-        "[end-para:::True]",
-        "[li(9,3):4::]",
-        "[para(9,5):]",
-        "[text(9,5):second list:]",
-        "[end-para:::True]",
-        "[block-quote(10,5)::> \n> ]",
-        "[para(10,7):]",
-        "[text(10,7):block within list:]",
-        "[end-para:::False]",
-        "[end-block-quote::> :True]",
-        "[tbreak(11,5):-::----]",
-        "[para(12,5):\n]",
-        "[text(12,5):first line\nanother list::\n]",
-        "[end-para:::True]",
-        "[li(14,4):5: :]",
-        "[para(14,6):]",
-        "[text(14,6):third list:]",
-        "[end-para:::True]",
-        "[block-quote(15,6)::> \n> ]",
-        "[para(15,8):]",
-        "[text(15,8):block within list:]",
-        "[end-para:::False]",
-        "[end-block-quote::> :True]",
-        "[tbreak(16,6):-::----]",
-        "[para(17,6):\n]",
-        "[text(17,6):first line\nanother list::\n]",
-        "[end-para:::True]",
-        "[end-ulist:::True]",
-        "[end-block-quote:::True]",
-        "[BLANK(19,1):]",
-    ]
-    expected_gfm = """<blockquote>
-<p>block start</p>
-<ul>
-<li>first list
-<blockquote>
-<p>block within list</p>
-</blockquote>
-<hr />
-first line
-another list</li>
-<li>second list
-<blockquote>
-<p>block within list</p>
-</blockquote>
-<hr />
-first line
-another list</li>
-<li>third list
-<blockquote>
-<p>block within list</p>
-</blockquote>
-<hr />
-first line
-another list</li>
-</ul>
-</blockquote>"""
-
-    # Act & Assert
-    act_and_assert(
-        source_markdown,
-        expected_gfm,
-        expected_tokens,
-        show_debug=False,
-    )
-
-
-@pytest.mark.gfm
 @pytest.mark.skip
 def test_extra_059g() -> None:
     """
@@ -18818,62 +17719,6 @@ another list</li>
         expected_gfm,
         expected_tokens,
         show_debug=False,
-    )
-
-
-@pytest.mark.gfm
-def test_extra_issue_1569_c() -> None:
-    """
-    TBD
-    """
-
-    # Arrange
-    source_markdown = """
-+ > first list
-  > + inner list
-  >  properly indented content
-  > + another list
-      properly indented content
-"""
-    expected_tokens = [
-        "[BLANK(1,1):]",
-        "[ulist(2,1):+::2::]",
-        "[block-quote(2,3):  :  > \n  > \n  > \n  > \n]",
-        "[para(2,5):]",
-        "[text(2,5):first list:]",
-        "[end-para:::True]",
-        "[ulist(3,5):+::6:: \n      ]",
-        "[para(3,7):\n]",
-        "[text(3,7):inner list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[li(5,5):6::]",
-        "[para(5,7):\n]",
-        "[text(5,7):another list\nproperly indented content::\n]",
-        "[end-para:::True]",
-        "[end-ulist:::True]",
-        "[end-block-quote:::True]",
-        "[BLANK(7,1):]",
-        "[end-ulist:::True]",
-    ]
-    expected_gfm = """<ul>
-<li>
-<blockquote>
-<p>first list</p>
-<ul>
-<li>inner list
-properly indented content</li>
-<li>another list
-properly indented content</li>
-</ul>
-</blockquote>
-</li>
-</ul>"""
-
-    # Act & Assert
-    act_and_assert(
-        source_markdown,
-        expected_gfm,
-        expected_tokens,
     )
